refactor(yaml): give the line-break rule a single home
`is_line_break` / `line_break_len` / `line_break_len_before` move verbatim out
of `light.rs` into a new `src/yaml/line_break.rs`, so the YAML 1.2 §5.4 rule —
`\n`, a lone `\r`, and `\r\n` as the two-byte spelling of *one* break — has one
definition that the oracle, the cursor, the validator and the newline index can
all share. Their names are unchanged, so the sixteen existing call sites in
`light.rs` need only the new import.

This is the module the rest of #341 routes into; nothing else changes yet.

Their unit tests move with them, and gain a cross-check that the three helpers
stay one rule seen from three angles: a byte is a break iff a break has non-zero
width there, and a break of width n starting at i is the break of width n ending
at i + n. The cross-check deliberately exempts positions inside a CRLF, where
the forward width is 1 (that LF) and the backward width is 2 (the whole CRLF) —
the two helpers answer different questions there, and `find_block_content_range`
depends on exactly that difference.

Refs #341, #324, #329, #106
